[Refector] seperate modules to files
